Make DataSerializable allow NSData. Fix comparison
Fixes critical bug where strncmp was used instead of memcmp. Ouch.

Also change `DataSerializable` to work in terms of a class method
instead of a failable initialiser. This way, a class cluster like
`NSData` can--and does--conform to `DataSerializable`. Just use `self`
in the class method to get an instance of type `Self`.

(Explanation: Using a failable initialiser would not work because
`NSData` has no `required` initialisers publicly available to be called
in the conforming `convenience init` of `DataSerializable`.)

Finally, fix the byte order of Swift integer instances of
`DataSerializable`.

Thank goodness we have unit tests now!